+++ R/CompDb-methods.R
@@ -213,3 +213,81 @@ setMethod("mass2mz", signature = c("ANY"),
           function(x, ...) {
               MetaboCoreUtils::mass2mz(x, ...)
           })
+
+#' @importFrom MsCoreUtils rbindFill
+#'
+#' @rdname CompDb
+#'
+#' @export
+setMethod("insertCompound", "CompDb", function(object, compounds = data.frame(),
+                                               addColumns = FALSE) {
+    if (!is.data.frame(compounds))
+        stop("'compounds' is expected to be a data.frame")
+    dbcon <- .dbconn(object)
+    if (is.null(dbcon)) stop("Database not initialized")
+    if (!nrow(compounds)) return(object)
+    ref <- data.frame(name = character(), inchi = character(),
+                      inchikey = character(), formula = character(),
+                      exactmass = numeric(), synonyms = character())
+    suppressWarnings(compounds <- rbindFill(compounds, ref))
+    compounds$compound_id <- as.character(compounds$compound_id)
+    .valid_compound(compounds, db = FALSE)
+    if (is.list(compounds$synonyms) | !all(is.na(compounds$synonyms))) {
+        if (is.list(compounds$synonyms)) {
+            syn <- data.frame(
+                compound_id = rep(compounds$compound_id,
+                                  lengths(compounds$synonyms)),
+                synonym = as.character(unlist(compounds$synonyms)))
+        } else syn <- data.frame(compound_id = compounds$compound_id,
+                                 synonym = compounds$synonyms)
+        syn <- syn[!is.na(syn$synonym), ]
+        if (nrow(syn)) dbAppendTable(dbcon, "synonym", syn)
+    }
+    dbcols <- colnames(dbGetQuery(dbcon, "select * from ms_compound limit 1"))
+    compounds$synonyms <- NULL
+    new_cols <- colnames(compounds)[!colnames(compounds) %in% dbcols]
+    if (addColumns && length(new_cols)) {
+        dtype <- dbDataType(dbcon, compounds[, new_cols, drop = FALSE])
+        dtype <- paste(names(dtype), dtype)
+        for (dt in dtype)
+            dbExecute(dbcon, paste("alter table ms_compound add", dt))
+        cols <- colnames(dbGetQuery(dbcon,
+                                    "select * from ms_compound limit 1"))
+        object@.properties$tables$ms_compound <- cols
+        dbcols <- colnames(
+            dbGetQuery(dbcon, "select * from ms_compound limit 1"))
+    }
+    dbAppendTable(dbcon, "ms_compound",
+                  compounds[, intersect(dbcols, colnames(compounds))])
+    object
+})
+
+#' @export
+#'
+#' @rdname CompDb
+setMethod("deleteCompound", "CompDb", function(object, ids = character(),
+                                               recursive = FALSE, ...) {
+    dbcon <- .dbconn(object)
+    if (is.null(dbcon)) stop("Database not initialized")
+    if (!length(ids)) return(object)
+    id_string <- paste0("'", ids, "'", collapse = ",")
+    if (hasMsMsSpectra(object)) {
+        ms <- dbGetQuery(dbcon, paste0("select compound_id, spectrum_id from ",
+                                       "msms_spectrum where compound_id in (",
+                                       id_string, ");"))
+        if (nrow(ms)) {
+            if (recursive)
+                object <- deleteSpectra(object, ids = ms$spectrum_id)
+            else stop("MS2 spectra for ", length(unique(ms$compound_id)),
+                      " of the specified compounds present. Use parameter ",
+                      "'recursive = TRUE' to delete compounds and all ",
+                      "related MS2 spectra from the database.")
+
+        }
+    }
+    dbExecute(dbcon, paste0("delete from synonym where compound_id in (",
+                            id_string, ");"))
+    dbExecute(dbcon, paste0("delete from ms_compound where compound_id in (",
+                            id_string, ");"))
+    object
+})